Update cash transfer to employee
Let me explain why I think this is wrong:

If we do cash transfer to owner it says a positive amount gets converted to Privatentnahme. For me, that is right. Because cash is getting less, so it is negative. Thus the other pay item needs to be positive.

Same goes for cash transfer to employee, no? If it is negative, then the cash pay item needs to be positive. That is wrong though because we give money to the employee, not take money from him.